requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2

fastapi
uvicorn[standard]
azure-monitor-opentelemetry


opentelemetry-api
opentelemetry-sdk
opentelemetry-semantic-conventions
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-distro
opentelemetry-exporter-otlp
#opentelemetry-bootstrap --action=install



#python -m uvicorn main:app --reload
#opentelemetry-instrument <your_run_command>